Moving test outside of cmd package since it will start the E2E tests
